[Editor] world saving
Added stub map to assets

To save the world, press 's' in editor mode

Closes #41

Also fixes two compilation errors

1) add implementation to GenericValue copy constructor
2) cast to long

+++ source/nuFileUtils.cpp
@@ -12,13 +12,13 @@ char* FileUtils::readFile(std::string fname)
 
     if (file.is_open())
     {
-	file.seekg(0, std::ios::end);
-	long size = file.tellg();
-	printf("readFile(): size=%ld\n", size);
-	buffer = new char[size];
-	file.seekg(0, std::ios::beg);
-	file.read(buffer, size);
-	file.close();
+        file.seekg(0, std::ios::end);
+        long size = (long)file.tellg();
+        printf("readFile(): size=%ld\n", size);
+        buffer = new char[size];
+        file.seekg(0, std::ios::beg);
+        file.read(buffer, size);
+        file.close();
     }
 
     return buffer;
